Refactor and improve map padding handling and animations
