+++ src/lib/components/map-details/AirportDetailsBody.svelte
@@ -1,0 +1,35 @@
+<script lang="ts">
+  import AirportFlightsCard from '$lib/components/airport-details/AirportFlightsCard.svelte';
+  import AirportStatsCard from '$lib/components/airport-details/AirportStatsCard.svelte';
+  import AirportTimeCard from '$lib/components/airport-details/AirportTimeCard.svelte';
+  import AirportWeatherCard from '$lib/components/airport-details/AirportWeatherCard.svelte';
+  import type { FlightData, prepareVisitedAirports } from '$lib/utils';
+
+  type VisitedAirport = ReturnType<typeof prepareVisitedAirports>[number];
+
+  let {
+    airport,
+    relatedFlights,
+    onShowDepartures,
+    onShowArrivals,
+  }: {
+    airport: VisitedAirport;
+    relatedFlights: FlightData[];
+    onShowDepartures: (flightId?: number) => void;
+    onShowArrivals: (flightId?: number) => void;
+  } = $props();
+</script>
+
+<AirportStatsCard
+  flights={relatedFlights}
+  airportId={airport.id}
+  airlineCount={airport.airlines.length}
+/>
+<AirportTimeCard tz={airport.tz} />
+<AirportWeatherCard icao={airport.icao} tz={airport.tz} lon={airport.lon} />
+<AirportFlightsCard
+  flights={relatedFlights}
+  airportId={airport.id}
+  onShowAllDepartures={onShowDepartures}
+  onShowAllArrivals={onShowArrivals}
+/>
