+++ src/lib/components/map-details/RouteDetailsActions.svelte
@@ -1,0 +1,32 @@
+<script lang="ts">
+  import { Funnel, Locate, X } from '@o7/icon/lucide';
+
+  import { closeMapDetails, focusMapDetails } from '$lib/state.svelte';
+
+  import DetailsActionButton from './DetailsActionButton.svelte';
+
+  let {
+    hasFilters,
+    filterActive,
+    onToggleFilter,
+  }: {
+    hasFilters: boolean;
+    filterActive: boolean;
+    onToggleFilter: () => void;
+  } = $props();
+</script>
+
+<DetailsActionButton
+  label="Locate route"
+  icon={Locate}
+  onclick={focusMapDetails}
+/>
+{#if hasFilters}
+  <DetailsActionButton
+    label={filterActive ? 'Clear route filter' : 'Filter map to route'}
+    icon={Funnel}
+    onclick={onToggleFilter}
+    pressed={filterActive}
+  />
+{/if}
+<DetailsActionButton label="Close details" icon={X} onclick={closeMapDetails} />
