+++ src/lib/components/map-details/index.ts
@@ -1,4 +1,2 @@
 export { default as MapDetailsPane } from './MapDetailsPane.svelte';
 export { default as MapDetailsFrame } from './MapDetailsFrame.svelte';
-export { default as AirportDetailsContent } from './AirportDetailsContent.svelte';
-export { default as RouteDetailsContent } from './RouteDetailsContent.svelte';
